+++ cameraview/src/main/java/com/otaliastudios/cameraview/filter/BaseFilter.java
@@ -5,7 +5,6 @@ import android.opengl.GLES20;
 import androidx.annotation.NonNull;
 
 import com.otaliastudios.cameraview.CameraLogger;
-import com.otaliastudios.cameraview.filters.NoFilter;
 import com.otaliastudios.cameraview.internal.GlUtils;
 import com.otaliastudios.cameraview.size.Size;
