+++ cameraview/src/main/java/com/otaliastudios/cameraview/filter/Filter.java
@@ -1,7 +1,6 @@
 package com.otaliastudios.cameraview.filter;
 
 import com.otaliastudios.cameraview.CameraView;
-import com.otaliastudios.cameraview.filters.NoFilter;
 
 import androidx.annotation.NonNull;
