+++ cameraview/src/main/java/com/otaliastudios/cameraview/filter/NoFilter.java
@@ -1,4 +1,4 @@
-package com.otaliastudios.cameraview.filters;
+package com.otaliastudios.cameraview.filter;
 
 import androidx.annotation.NonNull;
